requirements.txt
python-dotenv
typing
langchain
langchain-text-splitters
langchain-core
langchain-community
langchain-openai
langchain-chroma
langgraph
pypdf